# api/openapi.yml
openapi: 3.0.3
info:
  title: RASP Service API
  description: API для управления расписанием занятий, бронированиями и посещаемостью
  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - REQUEST_FAILED
                - FAILED_TO_DECODE
                - NOT_FOUND
                - LOCKED
                - CONFLICT
                - SLOT_NOT_AVAILABLE
            message:
              type: string
      example:
        error:
          code: NOT_FOUND
          message: resource not found
...
    AvailabilityTemplateRequest:
      type: object
      required:
        - teacher_id
        - recurrence
        - slot_duration_minutes
        - start_date
        - end_date
        - enabled
      properties:
        teacher_id:
          type: string
          description: Идентификатор преподавателя
        recurrence:
          $ref: '#/components/schemas/RecurrenceConfig'
        slot_duration_minutes:
          type: integer
          minimum: 1
          description: Длительность слота в минутах
        start_date:
          type: string
          format: date
          example: "2024-01-01"
          description: Дата начала действия шаблона
        end_date:
          type: string
          format: date
          example: "2024-12-31"
          description: Дата окончания действия шаблона
        enabled:
          type: boolean
          description: Включен ли шаблон
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Уникальный идентификатор ресурса
...
paths:
  /availability_templates:
    post:
      tags:
        - Availability Templates
      summary: Создать шаблон доступности
      description: Создает новый шаблон доступности для преподавателя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityTemplateRequest'
            example:
              teacher_id: "teacher-123"
              recurrence:
                days:
                  - monday
                  - wednesday
                  - friday
                start_time: "09:00"
                end_time: "18:00"
              slot_duration_minutes: 60
              start_date: "2024-01-01"
              end_date: "2024-12-31"
              enabled: true
      responses:
        '201':
          description: Шаблон успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/AvailabilityTemplateResponse'
              example:
                template:
                  id: "template-123"
                  teacher_id: "teacher-123"
                  recurrence:
                    days:
                      - monday
                      - wednesday
                      - friday
                    start_time: "09:00"
                    end_time: "18:00"
                  slot_duration_minutes: 60
                  start_date: "2024-01-01"
                  end_date: "2024-12-31"
                  enabled: true
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: failed to decode request
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to create availability template
  /availability_templates/{id}:
    get:
      tags:
        - Availability Templates
      summary: Получить шаблон доступности
      description: Возвращает информацию о шаблоне доступности по идентификатору
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Шаблон найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/AvailabilityTemplateResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Шаблон не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get availability template
    put:
      tags:
        - Availability Templates
      summary: Обновить шаблон доступности
      description: Обновляет существующий шаблон доступности
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityTemplateRequest'
      responses:
        '200':
          description: Шаблон успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: '#/components/schemas/AvailabilityTemplateResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Шаблон не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get availability template
    delete:
      tags:
        - Availability Templates
      summary: Удалить шаблон доступности
      description: Удаляет шаблон доступности
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Шаблон успешно удален
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Шаблон не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get availability template
  /time_blocks:
    post:
      tags:
        - Time Blocks
      summary: Создать блок времени
      description: Создает новый блок времени (отпуск, больничный и т.д.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeBlockRequest'
            example:
              teacher_id: "teacher-123"
              start: "2024-01-15T00:00:00Z"
              end: "2024-01-20T23:59:59Z"
              reason: "Отпуск"
              type: "vacation"
      responses:
        '201':
          description: Блок времени успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  time_block:
                    $ref: '#/components/schemas/TimeBlockResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: failed to decode request
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to create availability template
  /time_blocks/{id}:
    get:
      tags:
        - Time Blocks
      summary: Получить блок времени
      description: Возвращает информацию о блоке времени по идентификатору
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Блок времени найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  time_block:
                    $ref: '#/components/schemas/TimeBlockResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Блок времени не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get time block
    put:
      tags:
        - Time Blocks
      summary: Обновить блок времени
      description: Обновляет существующий блок времени
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeBlockRequest'
      responses:
        '200':
          description: Блок времени успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  time_block:
                    $ref: '#/components/schemas/TimeBlockResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: failed to decode request
        '404':
          description: Блок времени не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to update time block
    delete:
      tags:
        - Time Blocks
      summary: Удалить блок времени
      description: Удаляет блок времени
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Блок времени успешно удален
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Блок времени не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to delete time block
  /slots/{id}:
    get:
      tags:
        - Slots
      summary: Получить слот
      description: Возвращает информацию о слоте по идентификатору
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Слот найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  slot:
                    $ref: '#/components/schemas/SlotResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Слот не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get slot
  /slots/batch:
    get:
      tags:
        - Slots
      summary: Получить слоты по идентификаторам
      description: Возвращает список слотов по переданным идентификаторам
      parameters:
        - $ref: '#/components/parameters/IdsQuery'
      responses:
        '200':
          description: Список слотов
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/SlotResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get slots
  /slots/generate:
    post:
      tags:
        - Slots
      summary: Генерировать слоты
      description: Запускает асинхронную генерацию слотов на основе шаблонов доступности
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotGenerateRequest'
            example:
              template_id: "template-123"
              from: "2024-01-01"
              to: "2024-01-31"
      responses:
        '202':
          description: Генерация слотов запущена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotGenerateResponse'
              example:
                job_id: "job-123"
        '400':
          description: Неверный запрос (требуется template_id или teacher_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: template_id or teacher_id is required
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to generate slots
  /bookings:
    post:
      tags:
        - Bookings
      summary: Создать бронирование
      description: Создает новое бронирование слота для студента. Поддерживает идемпотентность через заголовок Idempotency-Key
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            example:
              slot_id: "slot-123"
              student_id: "student-456"
      responses:
        '201':
          description: Бронирование успешно создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: slot_id is required
        '404':
          description: Ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '409':
          description: Слот недоступен для бронирования
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: SLOT_NOT_AVAILABLE
                  message: slot is not available
        '423':
          description: Ресурс заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: LOCKED
                  message: resource is locked
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to create booking
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Получить бронирование
      description: Возвращает информацию о бронировании по идентификатору
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Бронирование найдено
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to get booking
    delete:
      tags:
        - Bookings
      summary: Удалить бронирование
      description: Удаляет бронирование
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Бронирование успешно удалено
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to delete booking
  /bookings/{id}/cancel:
    put:
      tags:
        - Bookings
      summary: Отменить бронирование
      description: Отменяет существующее бронирование
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Бронирование успешно отменено
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to cancel booking
  /bookings/reschedule:
    post:
      tags:
        - Bookings
      summary: Перенести бронирование
      description: Переносит существующее бронирование на другой слот
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRescheduleRequest'
            example:
              booking_id: "107e3a10-2bfa-4a34-a4ed-68e12ee66374"
              new_slot_id: "slot-456"
      responses:
        '200':
          description: Бронирование успешно перенесено
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Бронирование или слот не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Новый слот недоступен для бронирования
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: SLOT_NOT_AVAILABLE
                  message: slot is not available
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{id}/confirm:
    post:
      tags:
        - Bookings
      summary: Подтвердить бронирование
      description: Подтверждает существующее бронирование (меняет статус с pending на confirmed)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Бронирование успешно подтверждено
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAILED_TO_DECODE
                  message: id is required
        '404':
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: resource not found
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: REQUEST_FAILED
                  message: failed to confirm booking
...